Add systematic error logging for API debugging
Intercept HTTP errors from the Readeck API, as well as unhandled
exceptions, and log both the original Kobo request (filter for `ERROR_DUMP`)
and the failing Readeck API response with full context (filter for
`READECK_API_ERROR`).

The point is to keep the logs minimal (no extensive debugging on HTTP 200)
while including enough information to reproduce and fix edge cases or bugs
in the proxy.

Note that API tokens are anonymized.